+++ clippy_lints/src/lib.rs
@@ -357,6 +357,7 @@ pub fn register_plugins(reg: &mut rustc_plugin::Registry) {
         methods::SEARCH_IS_SOME,
         methods::SHOULD_IMPLEMENT_TRAIT,
         methods::SINGLE_CHAR_PATTERN,
+        methods::SLICE_ITER_NTH,
         methods::TEMPORARY_CSTRING_AS_PTR,
         methods::WRONG_SELF_CONVENTION,
         minmax::MIN_MAX,