+++ clippy_lints/src/shadow.rs
@@ -66,7 +66,7 @@ fn check_fn(cx: &LateContext, decl: &FnDecl, block: &Block) {
     let mut bindings = Vec::new();
     for arg in &decl.inputs {
         if let PatKind::Binding(_, ident, _) = arg.pat.node {
-            bindings.push((ident.node.unhygienize(), ident.span))
+            bindings.push((ident.node, ident.span))
         }
     }
     check_block(cx, block, &mut bindings);
@@ -120,7 +120,7 @@ fn check_pat(cx: &LateContext, pat: &Pat, init: &Option<&Expr>, span: Span, bind
     // TODO: match more stuff / destructuring
     match pat.node {
         PatKind::Binding(_, ref ident, ref inner) => {
-            let name = ident.node.unhygienize();
+            let name = ident.node;
             if is_binding(cx, pat) {
                 let mut new_binding = true;
                 for tup in bindings.iter_mut() {
@@ -139,7 +139,6 @@ fn check_pat(cx: &LateContext, pat: &Pat, init: &Option<&Expr>, span: Span, bind
                 check_pat(cx, p, init, span, bindings);
             }
         }
-        // PatEnum(Path, Option<Vec<P<Pat>>>),
         PatKind::Struct(_, ref pfields, _) => {
             if let Some(ref init_struct) = *init {
                 if let ExprStruct(_, ref efields, _) = init_struct.node {
@@ -327,7 +326,7 @@ fn is_self_shadow(name: Name, expr: &Expr) -> bool {
 }
 
 fn path_eq_name(name: Name, path: &Path) -> bool {
-    !path.global && path.segments.len() == 1 && path.segments[0].name.unhygienize() == name
+    !path.global && path.segments.len() == 1 && path.segments[0].name.as_str() == name.as_str()
 }
 
 struct ContainsSelf {
@@ -337,7 +336,7 @@ struct ContainsSelf {
 
 impl<'v> Visitor<'v> for ContainsSelf {
     fn visit_name(&mut self, _: Span, name: Name) {
-        if self.name == name.unhygienize() {
+        if self.name == name {
             self.result = true;
         }
     }